+++ src/banking_app/Transaction.java
@@ -1,7 +1,7 @@
 package banking_app;
 
 public class Transaction {
-    public boolean transfer(Account fromAccount, Account toAccount, int amount) {
+    public boolean transfer(Account fromAccount, Account toAccount, double amount) {
         if (fromAccount.withdraw(amount)) {
             toAccount.deposit(amount);
             return true;
